Add responsive mobile navigation
